Handling form data with a single formData state
I now learned to handle multiple form values using a single state such as const [formData, setFormData]

Instead of manually creating a state for each field from the form. It is better to create a single formData state for a cleaner and elegant look

On the other hand, I learned to use spread operator.
  - A spread operator allows us to iterate over an iterable object such as an array or string